# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Internet Weather API",
    "description": "Internet Weather observes the public internet, detects meaningful change in the\nAI-engineering ecosystem, and explains it with traceable evidence.\n\nThe frontend consumes only this API. Provider APIs are never exposed directly.\n\nEvery data response carries a `freshness` object stating when the signals were\ncomputed and how many days of observation support them. Findings carry an\n`epistemic_status` of `observation`, `inference`, `hypothesis` or `unknown`, so\nspeculation is never presented as fact.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/trends": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "The Trends page, in one call",
        "description": "Bounded by design: four ranked bands plus recent events.\n\nA single request so the page paints at once \u2014 assembling one screen from four\nround trips is what makes an interface feel like a dashboard.",
        "operationId": "trends_api_trends_get",
        "parameters": [
          {
            "name": "subdomain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Subdomain"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subdomain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trends"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technologies": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Technology universe",
        "operationId": "technologies_api_technologies_get",
        "parameters": [
          {
            "name": "subdomain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Subdomain"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subdomain"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherState"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(rank|heating|cooling|anomaly|stars|name)$",
              "default": "rank",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnologyList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technologies/{slug}": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Research page header",
        "operationId": "technology_api_technologies__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
              "examples": [
                "mcp"
              ],
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnologyDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technologies/{slug}/history": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Signal history",
        "operationId": "technology_history_api_technologies__slug__history_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
              "examples": [
                "mcp"
              ],
              "title": "Slug"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 90,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnologyHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technologies/{slug}/relationships": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Related technologies",
        "operationId": "technology_relationships_api_technologies__slug__relationships_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
              "examples": [
                "mcp"
              ],
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnologyRelationships"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Detected ecosystem events",
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "name": "subdomain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Subdomain"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subdomain"
            }
          },
          {
            "name": "technology",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80,
                  "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "mcp"
              ],
              "title": "Technology"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventType"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Event with evidence",
        "operationId": "event_api_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "occurred_on": {
            "type": "string",
            "format": "date",
            "title": "Occurred On"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "magnitude": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Magnitude"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "epistemic_status": {
            "$ref": "#/components/schemas/EpistemicStatus"
          },
          "technology_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Slug"
          },
          "technology_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Name"
          },
          "repository_full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Full Name"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "detected_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected At"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLink"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "occurred_on",
          "title",
          "magnitude",
          "confidence",
          "epistemic_status"
        ],
        "title": "EventDetail"
      },
      "EventList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "freshness"
        ],
        "title": "EventList"
      },
      "EventSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "occurred_on": {
            "type": "string",
            "format": "date",
            "title": "Occurred On"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "magnitude": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Magnitude"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "epistemic_status": {
            "$ref": "#/components/schemas/EpistemicStatus"
          },
          "technology_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Slug"
          },
          "technology_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Name"
          },
          "repository_full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Full Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "occurred_on",
          "title",
          "magnitude",
          "confidence",
          "epistemic_status"
        ],
        "title": "EventSummary"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "release",
          "star_spike",
          "anomaly",
          "new_repository",
          "archived",
          "renamed"
        ],
        "title": "EventType",
        "description": "Discrete ecosystem events."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Subdomain": {
        "type": "string",
        "enum": [
          "agentic_ai",
          "llm_ecosystem",
          "rag",
          "ai_infra",
          "multimodal",
          "mlops",
          "ai_security"
        ],
        "title": "Subdomain",
        "description": "Technology subdomains within the AI-engineering domain."
      },
...
      "TechnologyCard": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subdomain": {
            "$ref": "#/components/schemas/Subdomain"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "headline": {
            "type": "boolean",
            "title": "Headline",
            "default": false
          },
          "weather_state": {
            "$ref": "#/components/schemas/WeatherState"
          },
          "signals": {
            "$ref": "#/components/schemas/SignalSnapshot"
          },
          "spark": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Spark"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "epistemic_status": {
            "$ref": "#/components/schemas/EpistemicStatus",
            "default": "observation"
          },
          "rank_overall": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Overall"
          },
          "rank_subdomain": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Subdomain"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "subdomain",
          "weather_state",
          "signals"
        ],
        "title": "TechnologyCard",
        "description": "One technology as it appears on Trends, including its hover preview.\n\nDeliberately self-sufficient: the hover intelligence preview must not fire a\nsecond request, so the sparkline and explanation ship with the card."
      },
      "TechnologyDetail": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subdomain": {
            "$ref": "#/components/schemas/Subdomain"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "headline": {
            "type": "boolean",
            "title": "Headline",
            "default": false
          },
          "first_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen At"
          },
          "weather_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeatherState"
              },
              {
                "type": "null"
              }
            ]
          },
          "signals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "epistemic_status": {
            "$ref": "#/components/schemas/EpistemicStatus",
            "default": "observation"
          },
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/RepositorySensor"
            },
            "type": "array",
            "title": "Repositories"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "subdomain",
          "freshness"
        ],
        "title": "TechnologyDetail",
        "description": "The Research page header: what we know, and what it rests on."
      },
      "TechnologyHistory": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/HistoryPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "points",
          "freshness"
        ],
        "title": "TechnologyHistory"
      },
      "TechnologyList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TechnologyCard"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "freshness"
        ],
        "title": "TechnologyList"
      },
      "TechnologyRelationships": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "related": {
            "items": {
              "$ref": "#/components/schemas/RelatedTechnology"
            },
            "type": "array",
            "title": "Related"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "related",
          "freshness"
        ],
        "title": "TechnologyRelationships"
      },
      "Trends": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "overview": {
            "$ref": "#/components/schemas/WeatherOverview"
          },
          "heating": {
            "items": {
              "$ref": "#/components/schemas/TechnologyCard"
            },
            "type": "array",
            "title": "Heating"
          },
          "cooling": {
            "items": {
              "$ref": "#/components/schemas/TechnologyCard"
            },
            "type": "array",
            "title": "Cooling"
          },
          "emerging": {
            "items": {
              "$ref": "#/components/schemas/TechnologyCard"
            },
            "type": "array",
            "title": "Emerging"
          },
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/TechnologyCard"
            },
            "type": "array",
            "title": "Anomalies"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "type": "array",
            "title": "Events"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          }
        },
        "type": "object",
        "required": [
          "overview",
          "freshness"
        ],
        "title": "Trends",
        "description": "Everything the Trends page renders."
      },
...
      "WeatherOverview": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "technology_count": {
            "type": "integer",
            "title": "Technology Count",
            "default": 0
          },
          "state_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/WeatherState"
            },
            "type": "object",
            "title": "State Counts"
          },
          "mean_momentum": {
            "type": "number",
            "title": "Mean Momentum",
            "default": 0.0
          },
          "subdomains": {
            "items": {
              "$ref": "#/components/schemas/SubdomainClimate"
            },
            "type": "array",
            "title": "Subdomains"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          }
        },
        "type": "object",
        "required": [
          "freshness"
        ],
        "title": "WeatherOverview",
        "description": "Global conditions. The one-glance answer."
      },
      "WeatherState": {
        "type": "string",
        "enum": [
          "hot",
          "emerging",
          "stable",
          "cooling",
          "breaking",
          "storm"
        ],
        "title": "WeatherState",
        "description": "Computed atmospheric state of a technology.\n\nNEVER assign these by hand. They are the output of\n``internetweather.analysis.weather_state.classify`` operating on measured\nsignals. A technology with insufficient history is STABLE with low\nconfidence \u2014 not EMERGING."
      },
...
    }
  },
...
}